% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{ctas_log}
\alias{ctas_log}
\title{Conditional app logger}
\usage{
ctas_log(...)
}
\arguments{
\item{...}{Passed to \code{\link[base:message]{base::message()}}.}
}
\description{
Prints a message to the console when the \code{ctasapp.verbose} option is \code{TRUE}
(the default). Set \code{options(ctasapp.verbose = FALSE)} or pass
\code{verbose = FALSE} to \code{\link[=run_ctas_app]{run_ctas_app()}} to silence all log output.
}
\keyword{internal}
