% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctasapp-pkg.R
\docType{package}
\name{ctasapp-package}
\alias{ctasapp}
\alias{ctasapp-package}
\title{ctasapp: Shiny App to Visualize ctas Output}
\description{
Interactive Shiny application for exploring results from the ctas (Clinical Timeseries Anomaly Spotter) R package. Provides drill-down visualizations of site-level outlier scores and subject-level time series data.
}
\author{
\strong{Maintainer}: Björn Koneswarakantha \email{bjoern.koneswarakantha@roche.com}

Authors:
\itemize{
  \item Björn Koneswarakantha \email{bjoern.koneswarakantha@roche.com}
}

}
\keyword{internal}
