% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_ctas.R
\name{filter_unscheduled}
\alias{filter_unscheduled}
\title{Filter out unscheduled visits}
\usage{
filter_unscheduled(df)
}
\arguments{
\item{df}{Data frame with a \code{timepoint_1_name} column.}
}
\value{
Filtered data frame.
}
\description{
Filter out unscheduled visits
}
\keyword{internal}
