% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{format_score}
\alias{format_score}
\title{Format outlier score for display}
\usage{
format_score(score)
}
\arguments{
\item{score}{Numeric vector.}
}
\value{
Character vector.
}
\description{
Format outlier score for display
}
\keyword{internal}
