% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_score_colors_plot}
\alias{get_score_colors_plot}
\title{Get score colours for plots from configuration}
\usage{
get_score_colors_plot()
}
\value{
Character vector of hex colour strings.
}
\description{
Get score colours for plots from configuration
}
