% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_score_colors_table}
\alias{get_score_colors_table}
\title{Get score colours for tables from configuration}
\usage{
get_score_colors_table()
}
\value{
Character vector of hex colour strings.
}
\description{
Get score colours for tables from configuration
}
