% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_score_colors_table_text}
\alias{get_score_colors_table_text}
\title{Get score text colours for tables from configuration}
\usage{
get_score_colors_table_text()
}
\value{
Character vector of hex colour strings.
}
\description{
Get score text colours for tables from configuration
}
