% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{plot_cat_bar_ci}
\alias{plot_cat_bar_ci}
\title{Bar chart with 95\% CI for categorical data with few timepoints}
\usage{
plot_cat_bar_ci(df_plot)
}
\arguments{
\item{df_plot}{Prepared data frame.}
}
\value{
A ggplot object.
}
\description{
Bar chart with 95\% CI for categorical data with few timepoints
}
\keyword{internal}
