% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{plot_cat_bar_stacked}
\alias{plot_cat_bar_stacked}
\title{Stacked bar chart for categorical data with many timepoints}
\usage{
plot_cat_bar_stacked(df_plot)
}
\arguments{
\item{df_plot}{Prepared data frame.}
}
\value{
A ggplot object.
}
\description{
Stacked bar chart for categorical data with many timepoints
}
\keyword{internal}
