% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_ctas_data}
\alias{sample_ctas_data}
\title{Sample ctas input data}
\format{
A list with four elements:
\describe{
\item{data}{Tibble with 6890 rows: subject_id, timepoint_rank,
timepoint_1_name, result, parameter_id, timepoint_2_name, baseline}
\item{subjects}{Tibble with 177 rows: subject_id, site, country, region}
\item{parameters}{Tibble with 2 rows: parameter_id, parameter_name,
parameter_category_1 through _3, and analysis config columns}
\item{queries}{Data frame of simulated clinical queries (approx 15\\%
of data points) generated by \code{\link[=simulate_query_data]{simulate_query_data()}}}
\item{pd}{Data frame of simulated protocol deviations (site-level
linkage) generated by \code{\link[=simulate_pd_data]{simulate_pd_data()}}}
}
}
\source{
Generated from \code{ctas::ctas_data} via \code{data-raw/sample_data.R}
}
\usage{
sample_ctas_data
}
\description{
A subset of \code{ctas::ctas_data} containing the input data, subjects, and
parameters tables used to generate sample results.
}
\keyword{datasets}
