% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{try_alluvial}
\alias{try_alluvial}
\title{Attempt to render an alluvial plot}
\usage{
try_alluvial(df_plot)
}
\arguments{
\item{df_plot}{Prepared data frame with val_cat, subject_id, site_label, timepoint_1_name.}
}
\value{
A ggplot if successful, NULL if it errors.
}
\description{
Attempt to render an alluvial plot
}
\keyword{internal}
