% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_DataInput.R
\name{mod_DataInput_ui}
\alias{mod_DataInput_ui}
\title{Data Input Module - UI}
\usage{
mod_DataInput_ui(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\description{
Provides a dropdown to select sample datasets or upload custom files.
Upload mode is selected by default. When uploading, 2 mandatory file
inputs (results, input) and 2 optional file inputs (untransformed,
queries) are shown with collapsible format documentation.
}
