% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_DataInput.R
\name{mod_DataInput_ui}
\alias{mod_DataInput_ui}
\title{Data Input Module - UI}
\usage{
mod_DataInput_ui(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\description{
Lets the user choose between the bundled ctas sample data and the
pharmaversesdtm-derived SDTM sample data. Upload support is reserved
for a future version.
}
