% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_DataInput.R
\name{mod_DataInput_ui}
\alias{mod_DataInput_ui}
\title{Data Input Module - UI}
\usage{
mod_DataInput_ui(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\description{
Placeholder data loader with a button to load bundled sample data.
Designed to be replaced with a custom data source module (e.g. snwflid).
}
