% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{plot_timeseries_panel}
\alias{plot_timeseries_panel}
\title{Single panel of timeseries plot for a group of sites}
\usage{
plot_timeseries_panel(df, thresh, sites)
}
\arguments{
\item{df}{Data frame, pre-filtered to one parameter with site-level score columns.}

\item{thresh}{Numeric threshold.}

\item{sites}{Character vector of sites to facet.}
}
\value{
A ggplot object.
}
\description{
Single panel of timeseries plot for a group of sites
}
\keyword{internal}
