% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_ctas_app}
\alias{run_ctas_app}
\title{Run the ctas Shiny App}
\usage{
run_ctas_app(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}.}
}
\description{
Launches the interactive ctas visualization application.
}
