% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\docType{data}
\name{score_colors}
\alias{score_colors}
\alias{SCORE_BREAKS}
\alias{SCORE_COLORS_PLOT}
\alias{SCORE_COLORS_TABLE}
\title{Score colour constants}
\format{
An object of class \code{numeric} of length 4.

An object of class \code{character} of length 5.

An object of class \code{character} of length 5.
}
\usage{
SCORE_BREAKS

SCORE_COLORS_PLOT

SCORE_COLORS_TABLE
}
\description{
Breakpoints and colour vectors shared between DT table styling and
ggplot2 plot colouring. \code{SCORE_COLORS_PLOT} uses "royalblue" for
below-threshold; \code{SCORE_COLORS_TABLE} uses "white".
}
\keyword{datasets}
